🌱 Split MissionSidebarContainer: extract actions hook
Moves all callbacks, deep-link wiring and layout side effects out of
MissionSidebarContainer into useMissionSidebarActions.ts.

Refs #21874
